saved

+++ core/interpreter/compiler/src/blocks_compiler.c
@@ -30,7 +30,7 @@ int blocks_compiler(ProgramCode_t *code, block_list_ptr blocks) {
       current_block = current_block->next;
       continue;
     }
-    for (int instr_iterator = 0; instr_iterator < current_block->instr_array->ir_size;
+    for (size_t instr_iterator = 0; instr_iterator < current_block->instr_array->ir_size;
          instr_iterator++) {
       if (current_block->instr_array->ir_array[instr_iterator].opcode == OP_JUMP) {
         if (!bc_add_instr(
